+++ src/main/java/com/swart/runwith/domain/course_post/mapper/CoursePostDtoMapper.java
@@ -3,7 +3,7 @@ package com.swart.runwith.domain.course_post.mapper;
 import static org.mapstruct.MappingConstants.ComponentModel.SPRING;
 
 import com.swart.runwith.domain.course_post.dto.controller.CoursePostCreateControllerRequestDto;
-import com.swart.runwith.domain.course_post.dto.service.CoursePostCreateServiceRequestDto;
+import com.swart.runwith.domain.course_post.dto.service.request.CoursePostCreateServiceRequestDto;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = SPRING)
