+++ src/main/java/com/swart/runwith/domain/course_post/service/CoursePostService.java
@@ -1,6 +1,6 @@
 package com.swart.runwith.domain.course_post.service;
 
-import com.swart.runwith.domain.course_post.dto.service.CoursePostCreateServiceRequestDto;
+import com.swart.runwith.domain.course_post.dto.service.request.CoursePostCreateServiceRequestDto;
 import com.swart.runwith.domain.course_post.dto.service.response.CoursePostReadServiceResponseDto;
 import java.util.List;
