+++ src/main/java/com/swart/runwith/domain/course_post/service/impl/CoursePostServiceImpl.java
@@ -1,6 +1,6 @@
 package com.swart.runwith.domain.course_post.service.impl;
 
-import com.swart.runwith.domain.course_post.dto.service.CoursePostCreateServiceRequestDto;
+import com.swart.runwith.domain.course_post.dto.service.request.CoursePostCreateServiceRequestDto;
 import com.swart.runwith.domain.course_post.dto.service.response.CoursePostReadServiceResponseDto;
 import com.swart.runwith.domain.course_post.entity.CoursePost;
 import com.swart.runwith.domain.course_post.exception.CoursePostErrorCode;
